some more tests moved that are not real unit tests
