+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/HelperMethodTests.cs
@@ -220,20 +220,6 @@ namespace com.csutil.tests {
             });
         }
 
-        [Fact]
-        public async Task TestTaskWithTimeout() {
-            var t = Log.MethodEntered("TestTaskWithTimeout Part 1");
-            var t1 = TaskV2.Delay(50);
-            await t1.WithTimeout(1000);
-            Log.MethodDone(t);
-            t = Log.MethodEntered("TestTaskWithTimeout Part 2");
-            await Assert.ThrowsAsync<TimeoutException>(async () => {
-                Task t2 = TaskV2.Delay(5000);
-                await t2.WithTimeout(50);
-                Log.MethodDone(t);
-            });
-        }
-
         private static async Task CreateAndRunATaskThatFails() {
             Task myFailedTask = null;
             await Assert.ThrowsAsync<AggregateException>(async () => {
